Added scripts to detect when the controller is pointing at a scatter plot point.
Used ray caster to project rays from the right controller to the points in the scatter plot. When the right trigger is pressed, a debug statement is logged. The scatter plot points have a collider component that detects when these rays are hit. Debug statements are added to detect collision.